requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
emoji==2.14.1
eval_type_backport==0.2.2
faiss-cpu==1.10.0
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
gotrue==2.11.4
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-deepseek==0.1.2
langchain-experimental==0.3.4
langchain-google-genai==2.0.11
langchain-mcp-adapters==0.0.2
langchain-openai==0.3.7
langchain-postgres==0.0.13
langchain-text-splitters==0.3.6
langdetect==1.0.9
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-cli==0.1.74
langgraph-sdk==0.1.53
langsmith==0.3.10
lxml==5.3.1
marshmallow==3.26.1
matplotlib==3.10.1
mcp==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.65.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfkit==1.0.0
pgvector==0.3.6
pillow==11.1.0
postgrest==0.19.3
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pyparsing==3.2.1
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
realtime==2.4.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sse-starlette==2.2.1
starlette==0.46.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.23
unstructured-client==0.30.6
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0